requirements.txt
streamlit
qrcode[pil]
opencv-python-headless
numpy
pillow